Do not use webapp eclass in vm-manager ebuild.
The reason is that webapp-config is using hardlinks by default when
installing a webapp. Since /var is on a separate volume on the default
installation, unpacking the stage4 tarball is broken.
We can not specify options for the webapp-config call in the ebuild (via
the webapp eclass) and we do not want to call webapp-config ourselves
since it makes things more complex.
Thus: do a direct install to /var/www/localhost for now.